fix(e2e): separate e2e tests into different suites
